backend/requirements.txt
polars>=1.10.0
xlsxwriter>=3.2.0
openpyxl>=3.1.2
fastexcel>=0.11.5
pandas>=2.2.0
google-genai>=1.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
strawberry-graphql[fastapi]>=0.257.0
sqlmodel>=0.0.22
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0
celery[redis]>=5.4.0
python-multipart>=0.0.9
aiohttp>=3.9.0
slowapi>=0.1.9
pytest>=8.0
pytest-asyncio>=0.23
